changing data type to match migration

+++ server/models/Entry.js
@@ -6,7 +6,7 @@ module.exports = (sequelize, DataTypes) => {
       defaultValue: DataTypes.UUIDV4
     },
     currentStock: DataTypes.REAL,
-    orderQuantity: DataTypes.REAL,
+    orderQuantity: DataTypes.INTEGER,
     comment: DataTypes.STRING
   })
